move tets for util funcs
